3.4-django-testing/django_testing/requirements.txt
django~=4.1.2
djangorestframework~=3.14.0
django-filter~=22.1
psycopg2-binary

pytest~=7.2.0